requirements.txt
ccxt>=4.0
pandas>=2.0
numpy>=1.24
vectorbt>=0.26
python-dotenv>=1.0
PyYAML>=6.0
APScheduler>=3.10
python-telegram-bot>=20.0
lightgbm>=4.0
optuna>=3.0
scikit-learn>=1.3
